--- src/automations/runner.ts
@@ -1,51 +1,0 @@
-import { DatabaseSingleton } from '@/db/singleton'
-import { chatMessagesTable, chatThreadsTable, modelsTable, promptsTable } from '@/db/tables'
-import { convertUIMessageToDbChatMessage } from '@/lib/utils'
-import type { UIMessage } from 'ai'
-import { and, eq, isNull } from 'drizzle-orm'
-import { v7 as uuidv7 } from 'uuid'
-
-type Navigate = (url: string) => void
-
-/**
- * Runs an automation by creating a new chat thread and seeding it with the prompt.
- * Navigation to the thread will trigger the AI streaming response.
- */
-export const runAutomation = async (promptId: string, navigate?: Navigate) => {
-  const db = DatabaseSingleton.instance.db
-
-  const prompt = await db
-    .select()
-    .from(promptsTable)
-    .where(and(eq(promptsTable.id, promptId), isNull(promptsTable.deletedAt)))
-    .get()
-  if (!prompt) throw new Error('Prompt not found')
-
-  const model = await db
-    .select()
-    .from(modelsTable)
-    .where(and(eq(modelsTable.id, prompt.modelId), isNull(modelsTable.deletedAt)))
-    .get()
-  if (!model) throw new Error('Model not found')
-
-  const threadId = uuidv7()
-
-  await db.insert(chatThreadsTable).values({
-    id: threadId,
-    title: prompt.title ?? 'Automation',
-    triggeredBy: prompt.id,
-    wasTriggeredByAutomation: 1,
-  })
-
-  const userMessage: UIMessage = {
-    id: uuidv7(),
-    role: 'user',
-    metadata: { modelId: model.id },
-    parts: [{ type: 'text', text: prompt.prompt }],
-  }
-
-  // First message in the thread, so no parent
-  await db.insert(chatMessagesTable).values(convertUIMessageToDbChatMessage(userMessage, threadId, null))
-
-  navigate?.(`/chats/${threadId}`)
-}
